Fix location of mute button in group details view
