+++ app/test/chat_details/goldens/group_details_empty.linux.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:e16571a855b9e1aa73b414d05e4d5135371c145c907b5969dd53536568609073
-size 149056
+oid sha256:5fc3df7435123d556a2add92e616b4de5d055e202a90f4d5de91fb789613b91b
+size 149229
