+++ app/test/chat_details/goldens/group_details_members_overflow.linux.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:f71682bc5bee44dcf06d807fe25687fe859d6f94d49daea8cdb1ea1c2c77de49
-size 172050
+oid sha256:e3ff57cd42fb3bbcc54b965b508f9d02a1b4a7e29b3f4d876bdb664c898ad003
+size 171435
